# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Prism Time-Series API
  version: 1.0.0
  description: |
    Единый контракт для всех бэкендов стенда.
    Канонические операции — POST. GET оставлен для отладки.
    Хранилище выбирается `PRISM_STORAGE`.
    Ошибки всегда в конверте `{ "error": { "code", "message" } }`.
servers:
  - url: http://localhost:8081
    description: Go API
  - url: http://localhost:8082
    description: Python API
paths:
  /healthz:
    get:
      summary: Liveness
      responses:
        "200":
          description: OK
  /readyz:
    get:
      summary: Readiness (хранилище доступно)
      responses:
        "200":
          description: Ready
        "503":
          $ref: "#/components/responses/Error"
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: Exposition format
  /v1/meta:
    get:
      summary: Активный backend/storage и поддерживаемые операции
      responses:
        "200":
          description: Meta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
  /v1/points:
    post:
      summary: Записать пачку точек
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteRequest"
      responses:
        "200":
          description: Written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResponse"
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/query:
    post:
      summary: Range-query с даунсемплом (канонический метод)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    get:
      summary: Range-query через query string
      parameters:
        - $ref: "#/components/parameters/Metric"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Step"
        - $ref: "#/components/parameters/Agg"
        - $ref: "#/components/parameters/Labels"
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          $ref: "#/components/responses/Error"
  /v1/latest:
    post:
      summary: Последняя точка (канонический метод)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LatestRequest"
      responses:
        "200":
          description: Point
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Point"
        "404":
          $ref: "#/components/responses/Error"
        "400":
          $ref: "#/components/responses/Error"
    get:
      summary: Последняя точка через query string
      parameters:
        - $ref: "#/components/parameters/Metric"
        - $ref: "#/components/parameters/Labels"
      responses:
        "200":
          description: Point
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Point"
        "404":
          $ref: "#/components/responses/Error"
components:
  parameters:
    Metric:
      in: query
      name: metric
      required: true
      schema: { type: string }
    From:
      in: query
      name: from
      required: true
      schema: { type: string, format: date-time }
    To:
      in: query
      name: to
      required: true
      schema: { type: string, format: date-time }
    Step:
      in: query
      name: step
      schema: { type: string, default: 1m, examples: ["15s", "1m", "5m"] }
    Agg:
      in: query
      name: agg
      schema: { type: string, enum: [avg, min, max, sum, count], default: avg }
    Labels:
      in: query
      name: labels
      description: JSON-объект фильтров, например {"host":"dev-1"}
      schema: { type: string }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Point:
      type: object
      required: [ts, metric, value]
      properties:
        ts:
          type: string
          format: date-time
        metric:
          type: string
          examples: ["cpu.usage"]
        value:
          type: number
        labels:
          type: object
          additionalProperties:
            type: string
          default: {}
    WriteRequest:
      type: object
      required: [points]
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/Point"
    WriteResponse:
      type: object
      required: [written]
      properties:
        written: { type: integer }
    QueryRequest:
      type: object
      required: [metric, from, to]
      properties:
        metric: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        step: { type: string, default: 1m }
        agg: { type: string, enum: [avg, min, max, sum, count], default: avg }
        labels:
          type: object
          additionalProperties: { type: string }
          default: {}
    LatestRequest:
      type: object
      required: [metric]
      properties:
        metric: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
          default: {}
    QueryResult:
      type: object
      required: [metric, agg, step, points]
      properties:
        metric: { type: string }
        agg: { type: string }
        step: { type: string }
        points:
          type: array
          items:
            type: object
            required: [ts, value]
            properties:
              ts: { type: string, format: date-time }
              value: { type: number }
    Meta:
      type: object
      required: [backend, storage, storages, contract, ops]
      properties:
        backend: { type: string }
        storage: { type: string }
        storages:
          type: array
          items: { type: string }
        contract: { type: string, examples: ["v1"] }
        ops:
          type: array
          items: { type: string, enum: [write, query, latest] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, not_found, storage_unavailable, storage_error]
            message: { type: string }